fix: 로그인, 회원 등록 완료 모달 추
