+++ test-web/src/components/LogIn/LogInField.module.css
@@ -1,0 +1,13 @@
+:root {
+    --font-size-vw: calc((50 / 1920) * 100vw); /* Define a CSS variable */
+  }
+  
+  .customText {
+    color: #616161;
+    font-family: 'Google Sans';
+    font-size: var(--font-size-vw); /* Use the CSS variable */
+    font-style: normal;
+    font-weight: 400;
+    line-height: normal;
+  }
+
